+++ interface/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class InterfaceConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "interface"
+
+    def ready(self):
+        import interface.signals
